Merge

+++ web-app/src/app/component/expenses/expenses.component.ts
@@ -4,9 +4,7 @@ import { ExpenseService } from 'src/app/shared/expense.service';
 import { Subject, takeUntil, tap, timestamp } from 'rxjs';
 import { NgForm } from '@angular/forms';
 import { AngularFirestore } from '@angular/fire/compat/firestore';
-import { AngularFireAuth } from '@angular/fire/compat/auth';
-import firebase from 'firebase/compat/app';
-import { Timestamp } from 'firebase/firestore';
+import { Expenses } from 'src/app/model/expenses';
 
 @Component({
   selector: 'app-expenses',
@@ -18,18 +16,48 @@ export class ExpensesComponent {
   constructor(
     private auth: AuthService,
     private expenseService: ExpenseService,
-    private fireAuth: AngularFireAuth,
     private firestore: AngularFirestore
   ) {}
 
-  expenseData: any[] = [];
-  totalAmount: number = 0; // Initialize totalAmount
-  uid: string | null = null;
+  expenseData: Expenses[] = [];
+  totalAmount: number = 0;
+  uid: string | undefined;
 
   private unsubscribe$ = new Subject<void>();
   isLoading = false;
   isFilled = false;
   isEmpty = false;
+
+  selectedValue: string = '';
+  categories = [
+    { value: 'Food', viewValue: 'Food' },
+    { value: 'Entertainment', viewValue: 'Entertainment' },
+    { value: 'Grocery', viewValue: 'Grocery' },
+    { value: 'Shopping', viewValue: 'Shopping' },
+    { value: 'Bills', viewValue: 'Bills' },
+    { value: 'Transportation', viewValue: 'Transportation' },
+  ];
+
+  ngOnInit() {
+    this.uid = this.auth.getUserId();
+    console.log('User ID: ' + this.uid);
+    if (!this.uid) {
+      console.error('Not currently signed in');
+      return;
+    }
+    this.isFilled = true;
+    this.getExpense();
+  }
+
+  ngOnDestroy(): void {
+    console.log('DashboardComponent destroyed');
+    this.unsubscribe$.next();
+    this.unsubscribe$.complete();
+  }
+
+  logout() {
+    this.auth.logout();
+  }
 
   calculateTotalAmount() {
     this.totalAmount = this.expenseData.reduce(
@@ -40,21 +68,14 @@ export class ExpensesComponent {
 
   addExpense(form: NgForm) {
     this.isFilled = false;
-    console.log('current UID:', this.uid);
     if (!this.uid) {
-      console.error('Not currently signed in');
+      console.error('Cannot add expense, user ID not found');
       return;
     }
 
     const dateInt = new Date(form.value.date).getTime();
+    const date = new Date(form.value.date);
 
-    const firestoreTimestamp: Timestamp = Timestamp.fromDate(form.value.date);
-    console.log('Firestore Timestamp:', firestoreTimestamp);
-    const firestoreTimestampString = firestoreTimestamp.toDate().toString();
-
-    const date = new Date(firestoreTimestampString);
-
-    // Step 2: Format the date into "MMM DD YYYY" format
     const formattedDate = date.toLocaleDateString('en-US', {
       month: 'short',
       day: '2-digit',
@@ -63,10 +84,11 @@ export class ExpensesComponent {
 
     console.log(formattedDate); // Output: "Jun 06, 2024"
 
-    const expensesData: any = {
+    const expensesData: Expenses = {
       title: form.value.Title,
       amount: form.value.Amount,
       dateString: formattedDate,
+      category: this.selectedValue,
       dateInt: dateInt,
       date: date,
       description: form.value.description,
@@ -77,7 +99,8 @@ export class ExpensesComponent {
       !expensesData.title ||
       !expensesData.amount ||
       !expensesData.date ||
-      !expensesData.description
+      !expensesData.description ||
+      !expensesData.category
     ) {
       this.isFilled = false;
       console.error('All fields are required');
@@ -91,8 +114,6 @@ export class ExpensesComponent {
       .then((docRef) => {
         // Update the document with the generated ID
         expensesData.id = docRef.id;
-
-        // Update the document in Firestore with the generated ID
         form.reset();
       })
       .catch((error) => {
@@ -102,9 +123,8 @@ export class ExpensesComponent {
 
   getExpense() {
     this.isLoading = true;
-    console.log('User ID: ' + this.uid);
     if (!this.uid) {
-      console.error('Not currently signed ins');
+      console.error('Cannot get expenses, user ID not found');
       return;
     }
 
@@ -129,45 +149,25 @@ export class ExpensesComponent {
   }
 
   deleteExpense(id: string) {
+    if (!id) {
+      console.error('Cannot delete expense without ID');
+      return;
+    }
+
     this.firestore
       .collection('expenses')
       .doc(id)
       .delete()
       .then(() => {
         console.log('Document successfully deleted!');
+        // Remove from local array
+        this.expenseData = this.expenseData.filter(
+          (expense) => expense.id !== id
+        );
+        this.calculateTotalAmount(); // Recalculate total amount
       })
       .catch((error) => {
         console.error('Error removing document: ', error);
       });
-  }
-
-  async getUserId() {
-    try {
-      const user = await this.fireAuth.currentUser;
-      if (user) {
-        this.uid = user.uid;
-        console.log('User ID: ' + this.uid);
-      } else {
-        console.log('No user is currently signed in.');
-      }
-    } catch (error) {
-      console.error('Error getting current user: ', error);
-    }
-  }
-
-  async ngOnInit(): Promise<void> {
-    await this.getUserId();
-    this.isFilled = true;
-    this.getExpense();
-  }
-
-  ngOnDestroy(): void {
-    console.log('DashboardComponent destroyed');
-    this.unsubscribe$.next();
-    this.unsubscribe$.complete();
-  }
-
-  logout() {
-    this.auth.logout();
   }
 }
